pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "finance-chat"
version = "0.1.0"
description = "Personal Finance AI Chat Assistant"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",       # for testing FastAPI endpoints
    "ruff>=0.4.0",         # linting
]

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
py-modules = []